# backend/internal/handler/vcjob/tensorflow.go
package vcjob

import (
	"fmt"

	"github.com/gin-gonic/gin"
	v1 "k8s.io/api/core/v1"
	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
	"k8s.io/utils/ptr"
	batch "volcano.sh/apis/pkg/apis/batch/v1alpha1"
	bus "volcano.sh/apis/pkg/apis/bus/v1alpha1"

	"github.com/raids-lab/orbit/internal/resputil"
	"github.com/raids-lab/orbit/internal/util"
	"github.com/raids-lab/orbit/pkg/config"
	"github.com/raids-lab/orbit/pkg/utils"
	"github.com/raids-lab/orbit/pkg/vcqueue"
)

type (
	PortReq struct {
		Name string `json:"name"`
		Port int32  `json:"port"`
	}

	TaskReq struct {
		Name       string          `json:"name"`
		Replicas   int32           `json:"replicas"`
		Resource   v1.ResourceList `json:"resource"`
		Image      ImageBaseInfo   `json:"image"`
		Shell      *string         `json:"shell"`
		Command    *string         `json:"command"`
		WorkingDir *string         `json:"workingDir"`
		Ports      []PortReq       `json:"ports"`
	}

	CreateTensorflowReq struct {
		CreateJobCommon `json:",inline"`
		Tasks           []TaskReq `json:"tasks"`
	}
)

// CreateTrainingJob godoc
//
//	@Summary		Create a training job
//	@Description	Create a training job
//	@Tags			VolcanoJob
//	@Accept			json
//	@Produce		json
//	@Security		Bearer
//	@Param			CreateTrainingReq	body		any						true	"CreateTrainingReq"
//	@Success		200					{object}	resputil.Response[any]	"Success"
//	@Failure		400					{object}	resputil.Response[any]	"Request parameter error"
//	@Failure		500					{object}	resputil.Response[any]	"Other errors"
//	@Router			/v1/vcjobs/tensorflow [post]
func (mgr *VolcanojobMgr) CreateTensorflowJob(c *gin.Context) {
	token := util.GetToken(c)

	var req CreateTensorflowReq
	if err := c.ShouldBindJSON(&req); err != nil {
		resputil.BadRequestError(c, err.Error())
		return
	}
	scheduleType, err := req.validateScheduleOptions(false)
	if err != nil {
		resputil.BadRequestError(c, err.Error())
		return
	}
	if !mgr.preCheckCreateJob(c, token, scheduleType, false) {
		return
	}
	scheduleMetadata, err := mgr.resolveJobScheduleMetadata(c.Request.Context(), scheduleType)
	if err != nil {
		resputil.Error(c, err.Error(), resputil.ServiceError)
		return
	}

	jobResources := utils.CalculateReplicatedResources(
		req.Tasks,
		func(task TaskReq) v1.ResourceList {
			return task.Resource
		},
		func(task TaskReq) int32 {
			return task.Replicas
		},
	)

	if err := vcqueue.EnsureAccountQueueExists(c, mgr.client, token, token.AccountID); err != nil {
		resputil.Error(c, fmt.Sprintf("failed to ensure account queue exists: %v", err), resputil.NotSpecified)
		return
	}
	if err := vcqueue.EnsureUserQueueExists(c, mgr.client, token, token.AccountID, token.UserID); err != nil {
		resputil.Error(c, fmt.Sprintf("failed to ensure user queue exists: %v", err), resputil.NotSpecified)
		return
	}

	// 如果希望接受邮件，则需要确保邮箱已验证
	if req.AlertEnabled && !utils.CheckUserEmail(c, token.UserID) {
		resputil.Error(c, "Email not verified", resputil.UserEmailNotVerified)
		return
	}

	// Generate job name with type prefix (RFC 1035 compliant)
	jobName := utils.GenerateJobName("tf", token.Username)
	// baseURL for ingress paths (without type prefix)
	baseURL := jobName[3:] // Remove "tf-" prefix

	// 1. Volume Mounts
	volumes, volumeMounts, err := GenerateVolumeMounts(c, req.VolumeMounts, token)
	if err != nil {
		resputil.Error(c, err.Error(), resputil.NotSpecified)
		return
	}
	checkpoint, err := prepareCheckpointConfig(token, &req.CreateJobCommon, volumeMounts)
	if err != nil {
		resputil.Error(c, err.Error(), resputil.NotSpecified)
		return
	}

	// 2. Node Affinity and Tolerations
	baseAffinity := GenerateNodeAffinity(req.Selectors, jobResources)
	baseTolerations := GenerateTaintTolerationsForAccount(token)
	envs := AppendCheckpointEnvs(GenerateEnvs(c, token, req.Envs), checkpoint, jobName)

	// 3. Labels and Annotations
	labels, jobAnnotations, podAnnotations := getLabelAndAnnotations(
		OrbitJobTypeTensorflow,
		token,
		baseURL,
		&req.CreateJobCommon,
		scheduleMetadata,
	)
	if err := ApplyCheckpointAnnotations(jobAnnotations, req.Checkpoint); err != nil {
		resputil.Error(c, err.Error(), resputil.NotSpecified)
		return
	}

	// 4. Create the task spec
	tasks, minAvailable := buildTensorflowTasks(
		req.Tasks,
		baseAffinity,
		baseTolerations,
		volumes,
		volumeMounts,
		envs,
		labels,
		podAnnotations,
		req.CpuPinningEnabled,
	)

	queueName := vcqueue.ResolveJobQueueName(token)
	// 5. Create volcano job
	job := batch.Job{
		ObjectMeta: metav1.ObjectMeta{
			Name:        jobName,
			Namespace:   config.GetConfig().Namespaces.Job,
			Labels:      labels,
			Annotations: jobAnnotations,
		},
		Spec: batch.JobSpec{
			TTLSecondsAfterFinished: ptr.To(SevenDaySeconds),
			MinAvailable:            minAvailable,
			SchedulerName:           VolcanoSchedulerName,
			Plugins: map[string][]string{
				"env": {},
				"svc": {},
			},
			Policies: []batch.LifecyclePolicy{
				{
					Action: bus.RestartJobAction,
					Event:  bus.PodEvictedEvent,
				},
			},
			Queue: queueName,
			Tasks: tasks,
		},
	}

	if err = mgr.submitJob(c, token, &job); err != nil {
		resputil.Error(c, err.Error(), resputil.NotSpecified)
		return
	}

	resputil.Success(c, job)
}

func buildTensorflowTasks(
	inputTasks []TaskReq,
	baseAffinity *v1.Affinity,
	baseTolerations []v1.Toleration,
	volumes []v1.Volume,
	volumeMounts []v1.VolumeMount,
	envs []v1.EnvVar,
	labels map[string]string,
	podAnnotations map[string]string,
	cpuPinningEnabled bool,
) (tasks []batch.TaskSpec, minAvailable int32) {
	tasks = make([]batch.TaskSpec, len(inputTasks))
	minAvailable = int32(0)

	for i := range inputTasks {
		task := &inputTasks[i]
		taskAffinity := GenerateArchitectureNodeAffinity(task.Image, baseAffinity)

		ports := make([]v1.ContainerPort, len(task.Ports))
		for j, port := range task.Ports {
			ports[j] = v1.ContainerPort{
				ContainerPort: port.Port,
				Name:          port.Name,
				Protocol:      v1.ProtocolTCP,
			}
		}

		podSpec := generatePodSpecForParallelJob(task, taskAffinity, baseTolerations, volumes, volumeMounts, envs, ports, cpuPinningEnabled)
		taskSpec := batch.TaskSpec{
			Name:     task.Name,
			Replicas: task.Replicas,
			Template: v1.PodTemplateSpec{
				ObjectMeta: metav1.ObjectMeta{
					Labels:      labels,
					Annotations: podAnnotations,
				},
				Spec: podSpec,
			},
		}

		if task.Name == "worker" {
			taskSpec.Policies = []batch.LifecyclePolicy{
				{
					Action: bus.CompleteJobAction,
					Event:  bus.TaskCompletedEvent,
				},
			}
		}

		minAvailable += task.Replicas
		tasks[i] = taskSpec
	}

	return tasks, minAvailable
}